+++ x/staking/keeper/keeper_test.go
@@ -182,6 +182,13 @@ func getUBDKey(delAddr sdk.AccAddress, valAddr sdk.ValAddress) []byte {
 	return append(append(unbondingDelegationKey, addresstypes.MustLengthPrefix(delAddr)...), addresstypes.MustLengthPrefix(valAddr)...)
 }
 
+// getUBDByValIndexKey creates the index-key for an unbonding delegation, stored by validator-index
+// VALUE: none (key rearrangement used)
+func getUBDByValIndexKey(delAddr sdk.AccAddress, valAddr sdk.ValAddress) []byte {
+	unbondingDelegationByValIndexKey := []byte{0x33}
+	return append(append(unbondingDelegationByValIndexKey, addresstypes.MustLengthPrefix(valAddr)...), addresstypes.MustLengthPrefix(delAddr)...)
+}
+
 // getUnbondingDelegationTimeKey creates the prefix for all unbonding delegations from a delegator
 func getUnbondingDelegationTimeKey(timestamp time.Time) []byte {
 	bz := sdk.FormatTimeBytes(timestamp)
@@ -283,7 +290,7 @@ func (s *KeeperTestSuite) TestUnbondingDelegationsMigrationToColls() {
 			}
 			bz := stakingtypes.MustMarshalUBD(s.cdc, ubd)
 			s.ctx.KVStore(s.key).Set(getUBDKey(delAddrs[i], valAddrs[i]), bz)
-			s.ctx.KVStore(s.key).Set(stakingtypes.GetUBDByValIndexKey(delAddrs[i], valAddrs[i]), []byte{})
+			s.ctx.KVStore(s.key).Set(getUBDByValIndexKey(delAddrs[i], valAddrs[i]), []byte{})
 		},
 		"d03ca412f3f6849b5148a2ca49ac2555f65f90b7fab6a289575ed337f15c0f4b",
 	)